Abstracted deploy method in helper.go and use it in deploy.go and redeploy.go
